refactor: extract SearchDrawer component
